pyproject.toml
[tool.poetry]
name = "sys_user_mgmt_svc"
version = "1.0.0"
description = ""
authors = ["AIdev"]
readme = "README.md"
packages = [{include = "sys_user_mgmt_svc", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.poetry.scripts]
sys_user_mgmt_svc = "sys_user_mgmt_svc.main:main"

[tool.pytest.ini_options]
pythonpath = [ "src/" ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"